Supporting doubles better, SecondPassVisitor is starting to reach a breaking point in complexity

+++ src/main/java/lambda/primitives/Fn1DtoX.java
@@ -3,5 +3,5 @@ package lambda.primitives;
 import lambda.Fn1;
 
 @SuppressWarnings("serial")
-public abstract class Fn1DtoX<R> extends Fn1<Integer, R> {
+public abstract class Fn1DtoX<R> extends Fn1<Double, R> {
 }
